+++ src/main.rs
@@ -84,8 +84,8 @@ fn main() {
         if spawn_timer.ready {
             spawn_timer = Timer::from_millis(Uniform::new(1000, 5000).sample(&mut rng));
             let to_coord = Coord::new(
-                Uniform::new(1, world.floor.rows as u16).sample(&mut rng),
-                Uniform::new(1, world.floor.cols as u16).sample(&mut rng),
+                Uniform::new(1, world.floor.rows).sample(&mut rng),
+                Uniform::new(1, world.floor.cols).sample(&mut rng),
             );
             if to_coord != player.coord {
                 world.monsters.push(Monster::new(to_coord, &mut rng));
